+++ src/source/geojson_source.test.ts
@@ -1,3 +1,4 @@
+import type {StyleSpecification} from '@maplibre/maplibre-gl-style-spec';
 import {describe, test, expect, vi} from 'vitest';
 import {Tile} from './tile';
 import {OverscaledTileID} from './tile_id';
@@ -12,9 +13,11 @@ import {SubdivisionGranularitySetting} from '../render/subdivision_granularity_s
 import {type ActorMessage, MessageType} from '../util/actor_messages';
 import {type Actor} from '../util/actor';
 import {MercatorTransform} from '../geo/projection/mercator_transform';
-import {sleep, waitForEvent} from '../util/test/util';
+import {sleep, waitForEvent, beforeMapTest, createMap as globalCreateMap} from '../util/test/util';
 import {type MapSourceDataEvent} from '../ui/events';
+import {type Source} from './source';
 import {type GeoJSONSourceDiff} from './geojson_source_diff';
+import type {MapDataEvent} from '../ui/events';
 
 const wrapDispatcher = (dispatcher) => {
     return {
@@ -61,6 +64,18 @@ const hawkHill = {
         }
     }]
 } as GeoJSON.GeoJSON;
+
+type MapOptions = {
+    style: StyleSpecification;
+};
+
+function createMap(options: MapOptions) {
+    const container = window.document.createElement('div');
+    window.document.body.appendChild(container);
+    Object.defineProperty(container, 'clientWidth', {value: 512});
+    Object.defineProperty(container, 'clientHeight', {value: 512});
+    return globalCreateMap({container, ...options});
+}
 
 describe('GeoJSONSource.constructor', () => {
     const mapStub = {
@@ -454,7 +469,7 @@ describe('GeoJSONSource.update', () => {
         } as any;
 
         source.on('data', (e) => {
-            if (e.sourceDataType === 'metadata') {
+            if (e.sourceDataType === 'content') {
                 source.setData({} as GeoJSON.GeoJSON);
                 source.loadTile(new Tile(new OverscaledTileID(0, 0, 0, 0, 0), source.tileSize));
             }
@@ -489,6 +504,58 @@ describe('GeoJSONSource.getData', () => {
 });
 
 describe('GeoJSONSource.updateData', () => {
+    test('Emits sourceDataChanged event boolean when source data is changed', async () => {
+        beforeMapTest();
+        const map = createMap({
+            style: {
+                version: 8,
+                sources: {geojson: {type: 'geojson', data: {type: 'FeatureCollection', features: []}}},
+                layers: [{id: 'geojson', type: 'fill', source: 'geojson'}]
+            }
+        });
+        await map.once('load');
+        const source: GeoJSONSource = map.getSource('geojson');
+
+        let sourceDataChanged = false;
+        (source as Source).fire = (event: MapSourceDataEvent) => {
+            if (event.type === 'data' && event.sourceDataChanged) sourceDataChanged = true;
+        };
+
+        const update: GeoJSONSourceDiff = {
+            add: [{id: '1', type: 'Feature', properties: {}, geometry: {type: 'LineString', coordinates: []}}],
+        };
+        source.updateData(update);
+        await map.once('idle');
+
+        expect(sourceDataChanged).toBe(true);
+    });
+
+    test('Fires data event to update (render) map only once for a single diff update', async () => {
+        beforeMapTest();
+        const map = createMap({
+            style: {
+                version: 8,
+                sources: {geojson: {type: 'geojson', data: {type: 'FeatureCollection', features: []}}},
+                layers: [{id: 'geojson', type: 'fill', source: 'geojson'}]
+            }
+        });
+        await map.once('load');
+        const source: GeoJSONSource = map.getSource('geojson');
+
+        let fireCount = 0;
+        (source as Source).fire = (event: MapDataEvent) => {
+            if (event.type === 'data') fireCount++;
+        };
+        // add one feature
+        const update: GeoJSONSourceDiff = {
+            add: [{id: '1', type: 'Feature', properties: {}, geometry: {type: 'LineString', coordinates: []}}],
+        };
+        source.updateData(update);
+        await map.once('idle');
+
+        expect(fireCount).toBe(1);
+    });
+
     test('queues a second call to updateData', async () => {
         const spy = vi.fn();
         const mockDispatcher = wrapDispatcher({
@@ -504,7 +571,7 @@ describe('GeoJSONSource.updateData', () => {
 
         // Wait for initial data to be loaded
         source.load();
-        await waitForEvent(source, 'data', (e: MapSourceDataEvent) => e.sourceDataType === 'metadata');
+        await waitForEvent(source, 'data', (e: MapSourceDataEvent) => e.sourceDataType === 'content');
 
         spy.mockClear();
 
@@ -523,8 +590,8 @@ describe('GeoJSONSource.updateData', () => {
         source.updateData(update2);
 
         // Wait for both updateData calls to be performed
-        await waitForEvent(source, 'data', (e: MapSourceDataEvent) => e.sourceDataType === 'metadata');
-        await waitForEvent(source, 'data', (e: MapSourceDataEvent) => e.sourceDataType === 'metadata');
+        await waitForEvent(source, 'data', (e: MapSourceDataEvent) => e.sourceDataType === 'content');
+        await waitForEvent(source, 'data', (e: MapSourceDataEvent) => e.sourceDataType === 'content');
 
         expect(spy).toHaveBeenCalledTimes(2);
         expect(spy.mock.calls[0][0].data.dataDiff).toEqual(update1);
@@ -563,8 +630,8 @@ describe('GeoJSONSource.updateData', () => {
         source.updateData(update2);
 
         // Wait for the setData and updateData calls to be performed
-        await waitForEvent(source, 'data', (e: MapSourceDataEvent) => e.sourceDataType === 'metadata');
-        await waitForEvent(source, 'data', (e: MapSourceDataEvent) => e.sourceDataType === 'metadata');
+        await waitForEvent(source, 'data', (e: MapSourceDataEvent) => e.sourceDataType === 'content');
+        await waitForEvent(source, 'data', (e: MapSourceDataEvent) => e.sourceDataType === 'content');
 
         expect(spy).toHaveBeenCalledTimes(2);
         expect(spy.mock.calls[0][0].data.data).toEqual(JSON.stringify(data1));
@@ -605,8 +672,8 @@ describe('GeoJSONSource.updateData', () => {
         source.setData(data2);
 
         // Wait for both setData calls to be performed
-        await waitForEvent(source, 'data', (e: MapSourceDataEvent) => e.sourceDataType === 'metadata');
-        await waitForEvent(source, 'data', (e: MapSourceDataEvent) => e.sourceDataType === 'metadata');
+        await waitForEvent(source, 'data', (e: MapSourceDataEvent) => e.sourceDataType === 'content');
+        await waitForEvent(source, 'data', (e: MapSourceDataEvent) => e.sourceDataType === 'content');
 
         expect(spy).toHaveBeenCalledTimes(2);
         expect(spy.mock.calls[0][0].data.data).toEqual(JSON.stringify(data1));
@@ -643,9 +710,9 @@ describe('GeoJSONSource.updateData', () => {
         source.updateData(update1);
 
         // Wait for the calls to be performed
-        await waitForEvent(source, 'data', (e: MapSourceDataEvent) => e.sourceDataType === 'metadata');
-        await waitForEvent(source, 'data', (e: MapSourceDataEvent) => e.sourceDataType === 'metadata');
-        await waitForEvent(source, 'data', (e: MapSourceDataEvent) => e.sourceDataType === 'metadata');
+        await waitForEvent(source, 'data', (e: MapSourceDataEvent) => e.sourceDataType === 'content');
+        await waitForEvent(source, 'data', (e: MapSourceDataEvent) => e.sourceDataType === 'content');
+        await waitForEvent(source, 'data', (e: MapSourceDataEvent) => e.sourceDataType === 'content');
 
         expect(spy).toHaveBeenCalledTimes(3);
         expect(spy.mock.calls[0][0].data.data).toEqual(JSON.stringify(data1));
@@ -825,7 +892,7 @@ describe('GeoJSONSource.load', () => {
 
         // Wait for initial data to be loaded
         source.load();
-        await waitForEvent(source, 'data', (e: MapSourceDataEvent) => e.sourceDataType === 'metadata');
+        await waitForEvent(source, 'data', (e: MapSourceDataEvent) => e.sourceDataType === 'content');
 
         // Run again, with no additional data loaded
         source.load();