+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## main
 
 ### ✨ Features and improvements
+- Emit sourceDataChanged boolean and optimize GeoJSON events ([#6515](https://github.com/maplibre/maplibre-gl-js/pull/6515))
 - Improve fading - dynamic bi-directional raster cross-fading and self fading ([#6469](https://github.com/maplibre/maplibre-gl-js/pull/6469))
 - Support for usage of line-gradient together with line-dasharray
 - _...Add new stuff here..._